+++ src/lib.rs
@@ -56,8 +56,8 @@ pub struct Range {
     pub end_point: Point,
 }
 
+pub use crate::lint::Lint;
 pub use crate::lint::LintMatch;
-pub use crate::lint::LintMeta;
 pub use crate::lint::builtin_lints;
 pub use crate::lint::lint;
 pub use crate::report::Opts;