+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+Unreleased
+----------
+- Renamed `LintMeta` to `Lint` and made it exhaustive and user
+  constructible
+  - Added `message` member to it
+
+
 0.1.3
 -----
 - Added `core-read` lint